requirements.txt
spacy
numpy
pandas
re
mysql-connector-python
